Split PolynomialInMonomialBasis to its own file.
