+++ numerics/elliptic_integrals.cpp
@@ -11,7 +11,7 @@
 #include "glog/logging.h"
 #include "numerics/angle_reduction.hpp"
 #include "numerics/combinatorics.hpp"
-#include "numerics/polynomial.hpp"
+#include "numerics/polynomial_in_monomial_basis.hpp"
 #include "numerics/polynomial_evaluators.hpp"
 #include "quantities/elementary_functions.hpp"
 #include "quantities/si.hpp"
@@ -28,7 +28,7 @@ namespace internal {
 using namespace principia::base::_tags;
 using namespace principia::numerics::_angle_reduction;
 using namespace principia::numerics::_combinatorics;
-using namespace principia::numerics::_polynomial;
+using namespace principia::numerics::_polynomial_in_monomial_basis;
 using namespace principia::numerics::_polynomial_evaluators;
 using namespace principia::quantities::_elementary_functions;
 using namespace principia::quantities::_si;
