+++ numerics/newhall.hpp
@@ -6,6 +6,7 @@
 #include "base/not_null.hpp"
 #include "geometry/instant.hpp"
 #include "numerics/polynomial.hpp"
+#include "numerics/polynomial_in_monomial_basis.hpp"
 #include "numerics/чебышёв_series.hpp"
 #include "quantities/named_quantities.hpp"
 
@@ -17,6 +18,7 @@ namespace internal {
 using namespace principia::base::_not_null;
 using namespace principia::geometry::_instant;
 using namespace principia::numerics::_polynomial;
+using namespace principia::numerics::_polynomial_in_monomial_basis;
 using namespace principia::numerics::_чебышёв_series;
 using namespace principia::quantities::_named_quantities;
