+++ numerics/numerics.vcxproj
@@ -73,6 +73,8 @@
     <ClInclude Include="polynomial_body.hpp" />
     <ClInclude Include="polynomial_evaluators.hpp" />
     <ClInclude Include="polynomial_evaluators_body.hpp" />
+    <ClInclude Include="polynomial_in_monomial_basis.hpp" />
+    <ClInclude Include="polynomial_in_monomial_basis_body.hpp" />
     <ClInclude Include="quadrature.hpp" />
     <ClInclude Include="quadrature_body.hpp" />
     <ClInclude Include="root_finders.hpp" />
@@ -124,7 +126,7 @@
     <ClCompile Include="poisson_series_basis_test.cpp" />
     <ClCompile Include="poisson_series_test.cpp" />
     <ClCompile Include="polynomial_evaluators_test.cpp" />
-    <ClCompile Include="polynomial_test.cpp" />
+    <ClCompile Include="polynomial_in_monomial_basis_test.cpp" />
     <ClCompile Include="quadrature_test.cpp" />
     <ClCompile Include="root_finders_test.cpp" />
     <ClCompile Include="scale_b_test.cpp" />
